requirements.txt
notion-client
python-dotenv
streamlit
pandas
gspread
oauth2client
